requirements.txt
alembic==0.9.6
aniso8601==2.0.0
Babel==2.5.3
blinker==1.4
click==6.7
dominate==2.3.1
Flask==0.12.2
Flask-Admin==1.5.0
Flask-BabelEx==0.9.3
Flask-Bootstrap==3.3.7.1
Flask-HTTPAuth==3.2.3
Flask-Mail==0.9.1
Flask-Migrate==2.1.1
Flask-Moment==0.5.2
Flask-RESTful==0.3.6
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
Flask-Uploads==0.2.1
Flask-WTF==0.14.2
itsdangerous==0.24
Jinja2==2.10
Mako==1.0.7
MarkupSafe==1.0
Pillow==5.0.0
python-dateutil==2.6.1
python-editor==1.0.3
pytz==2017.3
six==1.11.0
speaklater==1.3
SQLAlchemy==1.1.15
visitor==0.1.3
Werkzeug==0.13
WTForms==2.1
